Created the loading of new history into the chat window and changes to different departments for communication
Created the loading of the new history of each department into the chat
window, while removing old history, also made it so that you can change
departments and it changes the channel appropriately to communicate only
to people in that channel you selected.  A lot of formatting changes
because of this as well.